Added More Status Items to Display
Per a request for enhancement, added more information for status to the display.  The display is extensible in that it can take more than just the temperature data, so that future devices can provide a more rich status.  The current SSD1306 display will now display status for the Fan (picture of a fan in the upper left), Igniter (picture of a flame in the middle left), Notifications (picture of a bell in the upper right), and Auger (picture of arrows in the middle right).  Also, the bottom of the screen displays the current mode in plain text.