Add ReIgnite Mode and Option to Reignite
Add ReIgnite Mode and Option to Reignite when temperature of the grill drops below minimum range.  You can attempt to reignite up to 10 times before a failure occurs.  Added to Safety Settings on the Settings page.   If using a previous version of PiFire, requires an update to your settings file to include the new setting.  This may mean that you will need to delete your settings file (or reset to factory settings) to get the new option setup.