Added STOP button after shutdown selected
Recently had an issue where i needed to stop and restart the grill quickly (my igniter died).  Unfortunately, the way it is setup is that you only have the option to go into the shutdown cycle.  I added the stop button, after you hit the shutdown to give you the option to stop immediately.  This allows you to restart immediately if necessary.