Minor Interface Visual Improvements / Status API
Added a few tweaks to the UI to improve the look of the notes in pages, modified the shutdown page so it has a spinner instead of a progress bar.  
Added API (/api/settings, /api/control, /api/current) which will only provide the status at the moment.  No changing the data right now.  FYI: I'm WIP creating a companion device that can be carried around the house to monitor the cook / change the temps, etc.