Updated History Chart AutoRefresh
Updated method of auto-refreshing the history chart to hopefully eliminate the issue causing the browser to run out of memory.  Instead of loading entirely new data every three seconds, using the chart.js API to update the data, and update the chart in place.  This also means that the interval of updates will depend on the number of datapoints and number of minutes selected to display in the settings page.