Adding Experimental Smoke-Plus Mode
This mode is highly experimental and has not been tested in real-world tests as of yet.  Please use with caution. 

Smoke-Plus mode will pulse the fan on / off (default 5 seconds on / 5 seconds off) during the Smoke mode, between default temperatures of 160F and 220F.  This is intended to promote the increased generation of smoke from the pellets in the firepot.  Default behavior, cycle time, minimum and maximum temperatures can be modified in the settings page.  

Note that default smoke mode is a simple fixed time cycle, with auger on/off times modulated by the P-Mode only.  It does not attempt to hold a specific temperature.  It's unclear how cycling the fan-speed will affect the pellet burn rates and pit temperatures.  Feedback is welcome.  

Note that settings.json format has been edited and will need to be rebuilt for existing installations.