Module Selection, Menu / Physical Button Inputs, Hopper Sensors Added
Some large changes to this commit.  First, a module selection script has been added to PiFire which means that it is easier to select different modules to be used with PiFire.  This has also been integrated into the setup/install script to install the right libraries for the current system.  

Launch the module selection at the command prompt by typing: 
$ sudo bash modules.sh 

The module selection will also affect how the flask app behaves.  If the prototype module is selected as the grill platform, it assumes you are running on a PC and launches in debug mode. 

Other changes: 
- Menu / Button Inputs:  Physical button inputs have been added to the SSD1306 display, which requires you to select the new SSD1306b display module.  A basic menu will pop up when any of the buttons is pressed and allows you to take basic actions from this menu.  It's rather crudely implemented, but it seems to work.  Needs more polish. 
- Hopper Level Sensor Modules:  Added modules to support hopper level sensing with the vl53l0x Time of Flight sensor and the HC-SR04 Ultrasonic Sensor (untested).  When using either of these modules, the dashboard will update with the hopper pellet level, as well as on status on the pellet manager page.  These pellet levels will update periodically (every 2.5m on the dashboard and every 1m on the pellet manager).  Note: The VL530L0X has been tested, but is definitely buggy.  The HC-SR04 has not been tested on real hardware.  

Note: 
- If you are updating from an existing install, please delete your settings.json and control.json files prior to running. 
- Reminder if installing from scratch, remember to use raspi-config to enable I2C for any devices that may utilize this bus (i.e. ADC, OLED)