Update to Dashboard JavaScript
Largely a behind the scenes update to overhaul the dashboard, especially around the control buttons on the bottom of the screen.  The intent of this change / overhaul was to improve performance and functionality of the buttons and how they update.  If you were paying close attention to previous version, these buttons would all refresh every 2-seconds which would result in the flickering of the buttons if hovering the mouse over them, and sometimes caused missed clicks for certain buttons.  This should be solved with the new JavaScript implementation, which only updates when the mode changes and is checked every two seconds, just as before.   

Moved dashboard to /dash and using redirect from / to redirect accesses to the root.  Renamed index.html to dash.html, introduced dash.js for the JavaScript support.  

Note: 
- If you are updating from an existing install, consider deleting your settings.json and control.json files prior to running. 
- Reminder if installing from scratch, remember to use raspi-config to enable I2C for any devices that may utilize this bus (i.e. ADC, OLED)