Add a Custom Grill Name!
I thought it would be fun to add a name for your beloved grill which is displayed in the navigation bar and on your Desktop PC in the browser title.  It might also be used on the display in the future.  

This could be a way to discern between your different grills if you have multiple PiFire smokers, you amazing person.  

Have fun with it! 

Note: 
- If you are updating from an existing install, please delete your settings.json prior to running.