Added ST7789P (Pimoroni) Display and PyGame Display Modules
Added support for the ST7789P (240x240 IPS/TFT) display using the Pimoroni python libraries.  Added the P to the name of this module to distinguish it from any future modules based on other libraries (i.e. Luma.LCD).  

Added support for a debug display library based on PyGame.  When running on a PC for development, the display is shown in a new window.  Helpful for quickly mocking up new displays.  

Updated control.py, the installer and the modules.sh scripts to include both new display modules.