Dash Notification Enhancements
Changed notification buttons so that they are updated dynamically.  If a temperature achieves the set point, the notification button will go back to a bell with a slash through it instead of staying highlighted.  This way you do not have to refresh the page to see if there are any current notifications.  

To clear notifications that are currently set, simply click the button and press cancel in the pop-up window.  

You may have to do a hard-refresh on the dash screen (CTRL + Click Refresh) to get the functionality working properly.