Updated Physical Button Input
Updated the physical button input, converting from interrupt driven input to polling input.  This works better than the previous implementation in that it does not get spurious button input randomly, it avoid bugs with the display driver, and will allow for more interesting input capability in future releases.  

Responsiveness may still be something that needs to be worked on.