Updated ReadSettings() Function
Updated the ReadSettings() function to check for differences between the default keys and the loaded settings.  If any keys or subkeys are added to the json structure, then it will be written back to the settings file immediately.  

This should help with update scenarios, where feature settings have been added.  

Also, increased the default safety temperature to 550F.