Updated Pellet Level Accuracy
Added the "full" value to the pellet sensor calculation so that the pellet level can be more accurately calculated in the user interfaces.  

WARNING:  This may cause a breaking change for existing installs.  You should either edit your pelletdb.json with the following "full" value manually or delete the pelletdb.json and let PiFire create a new one on restart.  

...."empty": 22, "full": 4, "woods"....