Added Backup / Restore for Settings / PelletDB and more...
In preparation for the merge of development features back into main, I've added a backup / restore capability (in the admin panel) for both the settings and the pelletdb JSON files.  There is a local backup directory /backups where files are stored locally, but files can also be downloaded and restored from remote backups (via upload).  

The readme.md has been updated to include newer pictures, updated links, a new video and a features list.  Again this is all in anticipation for merging development changes back into the main.  

The install.sh script has been updated to do a shallow clone of PiFire so that most users aren't downloading the entire history and bulk of the repository, and only getting the latest. 

This version requires flask v2.x due to a change in the file upload capability.  If your install is on an older version, use 'pip3 install --upgrade flask' to update to the latest. 

A couple of minor bugfixes are included.