Changed Check for Prototype Dist, Changed Conversion to INT for F to C
Changed the check for prototype distance sensor to enable randomness when the grillplatform is also a prototype.  Moved the settings check so that it is done once during control startup instead of on every level check.  

Also changed the units conversion when changing from F to C so that it does not yield floats in the settings.json file.  Will enforce INTs going forward for C settings.