Added Exception Handling, Error Messages to WebUI
If any of the modules fail to load properly, an error message will be presented on the UI and the system will fallback to prototype modules.  This prevents the scripts from outright crashing.  

Added error if redis-server is not up and running, although the control script may still crash if this is the case.  

Fixed a bug with the wizard where the grillplat module in settings had a typo.