Minor Change to Updater to Resolve Interpreter Warning
In my current version of Python a warning is thrown that '!=' should be used instead of 'is not' in this particular case.  Not sure if this is an issue with my version of the interpreter, but I am making this change as I believe there should not be any harm.