Introducing Smart Start!
The new Smart Start feature, samples the initial grill temperature when starting-up to smartly configure both the startup time (time for ignition) and the p-mode & auger-on time settings for smoke mode.  

When the weather is colder, it frequently takes longer to get the grill started, and needs a longer startup/ignition cycle.  But when the weather outside is warm/hot startup can be extremely short.   In addition, the P-Mode usually needs to be adjusted seasonally to account for the ambient temperatures.  Smart Start makes it simple by deciding how to configure your startup and smoke cycle timings based on the initial grill temperature reading (usually similar to the ambient temperature).  

Enabling Smart Start is easy - just go to the settings, enable it with the toggle, configure the profiles to your liking, and save.  You will never have to mess with your settings again, no matter the season.  

As with any new feature, there may be bugs.  So please submit issues if you find any.  Enjoy!