Add Get Remote Branches to Updater
Add the ability to get remote branches.  On a fresh install with a shallow clone, it's likely that there will only be the main branch listed on the update page.  Executing this command will pull all of the remote branches so that the user can choose a branch from the remote.  

This command can mess up a development environment, so it will be suppressed unless run on a raspberry pi.