Added Display Type: None
While doing the release, I realized that if the prototype display is selected for an error scenario, the control logs will be spammed with prototype display data.  This is not ideal for SD Cards, so I created a display type 'none' which does very minimal display.   This can be used if the system has no display at all, or in the case of a display module error.