Giant Overhaul of Display Modules
In order to improve performance with the various displays, the display modules have been re-architected to utilize threads for the display and for an input.  All existing modules have been updated and tested.

This update should improve performance, though there may be timing challenges/bugs that haven't been observed yet.  Please submit any issues you discover to GitHub.  

The ili9341_encoder.py has been renamed to ili9341em.py and is used to support user lausdahl's implementation when using another SPI device in the system.  If you are using this module, you will need to run the wizard again, or update the settings.json. 

I have noticed that there may be some issues after running the wizard, and resetting the server.  I suspect that the system will need a full reboot between display configuration changes, so I may change this to a reboot in the future.