Bug Fix: Only Clear Screen Once on Inactivity
Clear screen and screen hide commands were being continuously sent to the ILI9341 display when the display was inactive.  This may have been causing the screen to become unresponsive after some time.  Fixed the code to ensure that the screen is not being constantly cleared on each loop through the display thread.