Changes to History Data Storage
This is a future looking change in the way that the history data is stored.  Instead of storing as a line of data with no designations of inputs, store the data as a JSON object that allows for defining specific inputs.  This will allow additional inputs and data to be stored in the history timeline for future implementations.  

The only impact is a slight increase in memory usage, however it should not be significant for the raspberry pi zero.