New Feature: Cook Files
This is a new feature that is being actively developed to save the history data for your cook.  Each cook is automatically saved to the ./history folder when it's completed (i.e. 'Stop' mode).  The cook file has an extension of .pifire and is a zip file that is comprised of all of the data related to the cook.  These files can be managed and viewed from the history page.  You can edit some of the data, including the title of the cook, and the labels of the temperature inputs.  In addition comments can be added.  

This feature has yet to be tested on real PiFire hardware.  

Future Plans:
* Upload & Download Cook Files from a cloud service (i.e. Google Drive)
* Add photos/pictures to a cook file