Changed All 240x320 Display Modules to use Base Class
Updated all display modules that are based off of the 240x320 resolution to use the same base class (display_base_240x320.py).  This will allow editing of the graphics and functionality in one place, versus having to change 9+ files for different displays.  

This was done to simplify maintenance and allow for some new feature ideas in this class of displays.  

ILI9341 and PyGame based display modules have been tested and are working.  If you have an ST7789, please test and provide any feedback on bugs.