Feature: Restart / Reboot / Shutdown Auto Reload if Ready
Added a server_status flag to the API so that the shutdown.html page can poll the server to see if it is up and ready prior to redirecting to the dashboard.  

The shutdown.html page will redirect to the dashboard as soon as the server is back up and available again after reboot/restart/shutdown.  No more waiting around, or getting bad gateway messages.