Breaking Change for ADS1x15 Module
In messing around with the ADS1x15 modules today, I noticed that the current version was actually using the ADS1015 library instead of the ADS1115 library.  So I've broken this out into two different modules that people can try.  This means you will likely have to go through the configuration wizard again to change to one of these two modules if you are currently using the ADS1x15 module.  

This may help determine if you have a ADS1015 that was sold as an ADS1115.