New Feature: Lid Open Detection
New feature to detect whether the lid was opened and suspends the PID while in hold mode for a set amount of time.  

Method of operation: 
In HOLD mode, once the target temperature has been achieved if the temperature drops below the threshold temperature (percentage of the set temperature), the PID will be paused for a preset amount of time.  The auger will still operate at the minimum cycle time, and the fan will be enabled (if it had been cycling for smoke plus mode).  

This feature can be enabled and configured in settings. 

!Warning!  This feature should be considered experimental until it has been tested on hardware.