Fix Bug: Control.py Hang in Smoke Mode, Display Hang in Error Mode
Fix a bug that was introduced with the Lid Open feature, where in Smoke Mode, a variable is not initialized.  This would cause the control.py to hang.  Added a check for Hold mode before checking this variable to avoid this issue. 

Fix a bug where the display would hang after an error occurs.  Implemented a short delay (so the error message can be seen briefly), and then clear display like in stop mode.