New Feature: Recipe Mode
Details on the recipe mode architecture can be found here: https://github.com/nebhead/PiFire/discussions/141

Note: This feature is experimental, and all caution should be applied when working with this feature.  There is no WebUI for this feature yet, and it can only be launched through manual redis database commands.  This is being provided for development purposes only.  

This feature should not adversely affect normal operation of PiFire, but it is possible that bugs have been introduced.  Please submit bugs on github.com.  Enjoy!