New Feature: Extended Data
This feature can be enabled in the Settings > History section and should be enabled only for debug / testing scenarios.  This allows the capturing of more data during the work cycles for things like Cycle Ratio, Set Points, Status, etc... 

Data can be exported during a cook on the history page, or from the cookfile itself.  

This feature should not be enabled/disabled while the grill is active.