MAX31865 Support and Cleanup
Added a probes module to support the MAX31865.  This module hasn't been tested yet on hardware yet, and thus may have bugs.  This is a straight conversion of the existing module from PiFire to the new format.  

Did some cleanup by removing the old probe modules (adc_*), and adding some documentation to the new probes_* files to assist with development.