Updated Notifications, Display Prototype Update
This commit brings updates to support notifications and the notification features with the updated probe structures.  Work is still needed for the influxdb integration, as it still contains older data structures. 

Display Prototype was updated to use Python Curses, for a nicer display output.   

This version will work as a standalone control.py module with all prototype modules.  The webui has not been updated and will not work with control.py currently.