Updated History/Status Page & Settings Page
Updated the history/status chart page so that it functions with the latest advanced probe changes.  

Updated the settings page to allow the selection of profiles for the advanced probe changes.   Also moved the naming of probes to the settings page, which should globally set the display names of the probes to whatever the user wants.  

Note: There is still much that will be broken in the current version.  For example, the wizard still needs to be updated so that the user can select the right configuration of probes (ADCs, RTDs, etc.) for their system.  In addition, both cookfiles and recipe modes have not been fully updated to work.  Display modules (except for the prototype display) are not currently updated and will likely crash if configured.  Etc.