Updated Wizard for Advanced Probes
Major update to the Configuration Wizard adding a probe configuration screen.  The probe configuration screen allows the user to select the device(s) to use for probe input, and then to configure the probe sources from each of those devices.  This is a bit complex, with many corner cases (especially in the case of virtual probes), so there is likely to be bugs.  

Other minor fixes: 
* read_settings now has an option to 'init' settings which is run on loading of the app and control scripts.  This will overlay defaults with the existing setting json file to grab updates.  This was changed so that this operation isn't done every single time the settings file is read, saving time/resources.  
* Force wizard to run on version check.  If v1.4.x or below, then force the wizard to run again.  TODO, cleanup of settings files. 
* Fixed an issue with start-to-mode feature where it would crash in settings.  
* Fixed issue with dashboard redirecting to wrong html file after default settings applied.