Updated Upgrade Path
Added upgrade path in read_settings such that specific settings in settings.json are cleaned-up in the transition from older versions.  

Added distance_none module as default, for systems that do not have a distance sensor.  Use this instead of prototype.  

Added none icon for selecting display none or distance none in wizard.