More Cookfile Updates
Added CSV download for both the cookfile and for the current status/history page.  

Added raw_data file to the cookfile format v1.5, which contains the same data that is stored in the cook database, including extended data (if that setting is enabled).  

Added fixup for the upgrade cookfile path, such that raw_data is properly added.  

Other small bug fixes.