Update ALL Displays to Support Advanced Probes
Updated all remaining displays to support advanced probes.  

There is definitely some room for improvement on the displays, as always. 

For the 320x240 displays, only the first two food probes will be displayed at the moment.  Some work could be done to configure which probes should be displayed, or cycled through.