Minor Change to Wizard, Making ADS1115_Adafruit Default
Given the challenges that many have experienced with the original ADS1115 module, the ADS1115 Adafruit module is now set to default for wizard configurations.  You can still select the original ADS1115 module or any other probe input module that you want in the wizard, but this makes it more straight-forward.