Implementing Read/Modify/Write for control in work_cycle
Due to some potential race conditions, where commands are not being accepted, adding read/modify/write in the work_cycle function to mitigate the issue.  

This may not fully mitigate the issue, thus in the future it may be necessary to implement a new command queue, so that commands are not lost.