Implemented Basic Command Queue
Implemented a basic command queue, which should not affect the API.  The way this command queue works is that writes to control are intercepted and placed into a queue, where the control script can process the queue at certain points of the execution flow.  This means that writes should not be lost anymore.  More testing is necessary to ensure that it is achieving the goals, and is not introducing any regressions.