Fixed MAX31865 probes module.  Added Adafruit version.
I finally got my hands on the MAX31865 module from Adafruit and have been able to test it against the existing module.  Unfortunately, this took me a good long while to tweak the settings for the original module, to get right.  In addition, I have added the Adafruit module, as it seems this is the original code that the contributed module borrowed from.  The only differences between these two modules is that the Adafruit module uses circuitpython libraries to implement the SPI interface, and prevents using either the CE0 or CE1 pins on the Raspberry PI.  Whereas the native module uses built-in spidev and utilizes either CE0 or CE1 for the chip select pin.  Having both modules provides options for someone who may want to mix and match other SPI devices on the same system.