Recipe Mode Web UI Initial Implementation
This is the initial implementation of the web UI for recipe mode.  This incorporates an update to the control panel as well, as a new user interface for viewing, running, and editing recipes.  

There will be bugs, of course, so please use at your own risk at the moment.  Submit any issues to GitHub.  

Enjoy!