Plugabble Controller Configuration Changes
Added more functionality to the configuration capabilities of the pluggable controller feature that could be used by add-on controllers.  Documentation is in the /controller folder.  

Updated the Fuzzy Logic controller with some additional rules, member function changes.  After some minimal testing, this controller is NOT ready for usage.  Temperature swings are significant and I would not recommend trying this until it can be tuned further.