Initial Machine Learning Controller Implementation
Initial ML controller implementation with Scikit Learn.  This implementation is not yet usable as a controller, and needs much more training data and experimentation before it should be considered ready to use.  In addition, there is no definition in the controllers.json and thus cannot be selected from the settings page.