Added SocketIO Hack for Backward Compatibility w/Android App
Added a translation layer (hack) for SocketIO calls, to maintain compatibility with the Android application.  This change is not likely to be 100% compatible, due to the architectural differences from v1.3.5 to the current v1.5.0+ codebases.  Please help submit any bugs that you may find with this. 

This hack will be removed in the future, after the Android App is updated.