Bug Fixes: Dash Notifications
Feedback from contributor @weberbox, fixed:

1. In the probe setpoint model you should not be able to select keep warm and shutdown at the same time. 
Fixed:  JS Updated in dashboard_macro files to toggle the other checkbox to be off if both are selected.  

2. On the Primary probe should keep warm and shutdown be an option?
Fixed: Removed these options from the UI on the primary grill probe. 

3. When clearing a probe notification it does not seem to be actually clearing them only setting req to false maybe? This was an issue before as I noticed it with the Android app and had to make changes to pifire. The probe setpoint as well as keep warm and shutdown options should be reset back to 0 and false in control when clearing a setpoint/notification IMHO

Fixed: Shutdown, KeepWarm and Target temp are all set back to defaults/cleared when cancelling, or when the temperature is achieved.