Bug Fix: Next Step Recipe Mode
In certain cases, the 'goto next step' button in the control panel would only move to the next step after two clicks.  This is due to the fact that it was not properly forcing the next step, when the pause bit was set.  This was resolved in the JS file and now should work properly.