Adding Basic Process Monitor for Control
Added a basic process monitor for the control script.  While running, this process monitor will check a heartbeat to ensure that the system hasn't timed out / hung.  If a timeout occurs, the system will attempt to restart all processes.  This is for additional safety.  

Current timeout is 30s.  Not currently tested on actual hardware.