Fix logic in probes/base.py
Fix logic in probes/base.py to ensure voltage is between 0mv and voltage reference plus some guardband percentage.   For this commit, Vs * 1.01 is the guardband amount.  This is meant to give reasonable range of error, and avoid spamming the logs if the Vs isn't set correctly in the device settings.