Fixes for MCP9600 Probes Device
Fix a minor bug with the default value for the i2c bus address, as well as actually selecting the right i2c bus address during initialization of the device.

This device should still be considered experimental.  This device appears to require a very slow i2c bus, which can cause issues with any other devices on the i2c bus (such as an ADS1115 or the distance sensor.