Improve VL530L0X Resilience
Adding a thread to read distance on the VL530L0X device rather than stalling the control script to read distance.  Stalls reading the distance can cause the control script to stall and the control monitor to trigger a restart of the control script.  By adding this read to a thread, we will not block the control thread during normal operation.