Fix Freeze on Stop Mode, Minor UI Improvements
Fixed a freeze when transitioning to Stop mode from shutdown or from a remote command.  Store the dash configuration when shutting off the display so that it doesn't need to build it each time you return to the dashboard.

Slight improvements to the dash object alignments.

Kill the display loop on shutdown, restart or reboot.