Improved Updater Data and Capabilities
Only updates packages that need to be updated for the version that you are updating to.  Update manifest now contains the latest version information.   Allow for selecting reboot on update if needed.  Added general commands to update process in case there are other scripts that need to be run or system settings that need to be modified during an update.