Bug Fixes from v1.6.0-Dev Build 30
Introduced a bug in the last version that caused Jinja to crash when settings wasn't passed into the base.html script on some pages.  Added some fault tolerance to this so that it should not crash in these scenarios.  Also, changed Pillow to >= 9.2.0 for compatibility with the latest Luma core.