Add DSI_Touch Image for Wizard, Update Wizard Data
Added a new image for the DSI Touch Display for the Configuration Wizard.  Added a note in the Wizard for the DSI Touch display that the installation of dependencies can take more than 10 minutes to complete as a whole desktop manager must be installed.  Also driving a reboot on upgrade to v1.7.0 due to the changes in udev rules.