Improve Pellet DB Resiliance
Pellet DB can get corrupted if power is removed at the wrong time, thus it is helpful to build in some resilience to pellet db to prevent this kind of issue causing the system to crash.

Backup PelletDB automatically when loading new pellets.  If corruption is detected when loading Pellet DB, attempt to load from a backup or grab default pellet db.