Update README for 1.6.0 Release / Improve ADS1115 Selection
Updated the README for the upcoming v1.6.0 release.  Improved the ADS1115 selection / documentation slightly in the Configuration Wizard so that the Adafruit ADS1115 module is listed first and the older ADS1115 module is renamed to 'Legacy' and a strong warning is provided to only use this if the Adafruit module does not seem to work.