Add Startup Exit Temperature Feature
Added startup exit temperature settings/feature so that when a specific threshold temperature is achieved, the startup exit's and transitions to the next mode.   This ensures that the startup doesn't run any longer than required.

This feature is provided for both standard startup and for smart start.  Set this value to 0 to disable the feature.