Modal Dialog Warning if System Active when Config Wizard Launched
A modal dialog will pop-up to remind the user if the system is active while attempting to run the configuration wizard.  This was requested by discord user @Richie who pointed out that it can be a bit annoying to get through with the configuration wizard only to realize that the system is running and have it fail.