Update Dash Data & Add Hide/Show Cards
Refactored the dashboard data structure in settings to  enable easier updates via the API.

Added 'Upgrade' path for settings if the build makes any breaking changes.

Made the deep dictionary update function common.

Added a gear icon in the corner of both the default and basic dashboards to allow for dashboard specific settings.  Added the ability to show/hide all of the different cards on these dashboards.