Allow Manual Input of Hold and Notification Temps from Dash
Allow the manual input of Hold and Notification target temperatures from the dashboard (default and basic).  Simply click on the large temperature value in the dialog window and enter in a new value using your keyboard or phone touch screen.  Inputs should snap to valid values and in the case of F input for hold mode, 5 degree increments.