Fix DS18B20 Device Not Found Issue
As described in the documentation for w1thermsensor, the soft-pull up also may require the RPi.GPIO module to set the the pull-up while in use.  The module must also wait for the 1-wire devices to be discovered by the kernel before trying to access them, thus a delay was added on initialization after the GPIO is held high.