Refactored Base_Flex Display
Completely refactored the base_flex display and the object loading to improve the flexibility, i.e. to use with other display devices in the future.  This will allow easier customization of displays by using a JSON formatted file to define the layout using flex objects that can be configured for that display.