Add Prime on Startup Feature
Added 'Always Prime on Startup' feature which allows the user to have a default prime amount on startup.  Note that if the user selects startup again during priming that was initiated with this feature then priming will be cancelled and the grill will enter startup mode.  (feature initially suggested by Discord user @TSJ)