Timer WebUI Overhaul
Overhauled timer UI on the web interface to utilize the newly created API calls (Eventually may deprecate the /timer function) and improve the UX.

Timers also will automatically detect if they have started from another source and display in the web interface.

Also included in this update, attempted to address the issue on the mobile Web UI where a long Grill Name (on the NAV Bar) would push the navbar menu to the next line.