New Debug and Health Features
New debug items added in the admin page.  You can now download the 'settings.json' and 'control:general' database files directly from the debug card.

Added 'check_alive' command for the platform control modules.

When the dashboard is loaded it will check if the control process is alive and flag an error if the control process does not respond.  This may be helpful in debugging issues as they arise.