Fix Bugs with DS18B20 Probe Type
Fix some bugs with the DS18B20 probe type as well as a change to make it run its sensor data gathering in a thread so that it's non-blocking.  The DS18B20 probes that were tested would sometimes take up to 1 second to respond which would cause a large delay in the rest of the system.