Auto-Tuning Refinement
Refined the auto-tune to use only the last measured resistance value for the specified temperature for calculations.  This assumes that as the temperature settles, the probe being measured may be slightly slower to react and thus the last measured resistance at any temperature from the reference probe is used for final calculations.