Added ProtoFlex to Allow for Faster Prototyping of Displays
Added ProtoFlex display module for faster prototyping .  In addition, overhauled some of the configuration in the Wizard for displays so that you can store your configuration in settings as a block.  Updated the control to ensure that certain settings were being applied from the correct configuration block in settings.  And probably more that I've forgotten about...