Refactor of FlexObject Module
Broke out each FlexObject type to individual classes.  This was done in an effort to improve memory efficiency and to simplify/re-use much of the code.  Unfortunately, minimal memory impact to this change, but hopefully the code is a little simpler to understand.  This makes it a bit easier to add new FlexObject types and to edit the existing ones.