Added Dashboard Configuration Capability
Added dashboard configuration capability defined in the ./dashboard folder with a json file defining configuration for each.  This will get built on boot and loaded dynamically.

Default dashboard now has the option to set the maximum temperature to display for both the primary and the food gauges.  This can be configured by clicking on the gear icon.