Added Offline Server Detection & Configuration to Basic Dash
Added offline server detection from the default and basic dashboards that lets you know if you have lost contact with the server.  This means that the server has stopped responding to requests and is likely offline (either off the network, in an errored state, not running, or off).

Also added the ability to add configuration settings to the basic dashboard in the future.