Resilience and Safety Improvements
Improved resilience of the control script, especially when loading probe devices.  If a probe device fails to load, the probes that are associated with that device will be disabled.

Added a critical error flag to the control script.  This will be used to detect when a critical error is encountered, which will prevent further control of the system (i.e. changing modes, etc.)